feat(home): real, indexable Spanish homepage at /es
The existing Español feature was a client-side swap of 14 hero strings
only — everything below the hero stayed English and Spanish was invisible
to crawlers. This replaces it with a full Spanish edition of the homepage.

- Extract homepage data fetching into src/lib/home-data.ts (getHomeData)
  and the page body into <HomeView lang>. Both `/` and `/es` render the
  same component, so the two language editions can't silently diverge.
- src/lib/home-i18n.ts: full en/es dictionary for all homepage chrome
  (hero, collections, discover, blog, about, support, search, footer
  essay) + Spanish names for the known collections.
- New SSR route /es with Spanish <title>/description/OpenGraph and
  hreflang alternates (en/es/x-default) on both routes — indexable.
- Hero language switch is now real cross-route links (/ ↔ /es) instead
  of a localStorage-only swap; es-locale visitors landing on `/` get a
  dismissible "view in Spanish" suggestion after mount (no SSR branch on
  Accept-Language, so the edge-cached English HTML is untouched).
- Remove now-unused src/lib/hero-i18n.ts.

Scope: localizes the homepage chrome. Book titles/translations stay in
their stored language (the corpus is overwhelmingly translated to
English), so /es is an honest Spanish gateway, not a Spanish-content
promise. Site nav (SiteHeader) and the editorial-spread/showcase
components remain English for now.
