+++ src/app/es/page.tsx
@@ -1,0 +1,34 @@
+import type { Metadata } from 'next';
+import { getHomeData } from '@/lib/home-data';
+import HomeView from '@/components/home/HomeView';
+
+// Spanish-language edition of the homepage — a real, server-rendered, indexable
+// route (not a client string swap), sharing the same data + body as `/`.
+export const revalidate = 60;
+export const maxDuration = 60;
+
+export const metadata: Metadata = {
+  title: 'Source Library — La mayor biblioteca de fuentes antiguas traducidas con IA',
+  description:
+    'Source Library digitaliza y traduce textos antiguos para estudiosos, buscadores y sistemas de IA. Explora miles de fuentes primarias de alquimia, hermética, filosofía y ciencia.',
+  alternates: {
+    canonical: '/es',
+    languages: {
+      en: '/',
+      es: '/es',
+      'x-default': '/',
+    },
+  },
+  openGraph: {
+    title: 'Source Library — Fuentes antiguas traducidas con IA',
+    description:
+      'La mayor biblioteca de acceso abierto de fuentes primarias traducidas. Alquimia, hermética, filosofía y ciencia, accesibles para todos.',
+    locale: 'es_ES',
+    url: 'https://sourcelibrary.org/es',
+  },
+};
+
+export default async function HomePageEs() {
+  const data = await getHomeData();
+  return <HomeView data={data} lang="es" />;
+}
