+++ src/components/home/HomeView.tsx
@@ -1,0 +1,392 @@
+import Image from 'next/image';
+import Link from 'next/link';
+import HeroSection from '@/components/layout/HeroSection';
+import HomePageSchema from '@/components/seo/HomePageSchema';
+import EditorialSpread from '@/components/prototype/EditorialSpread';
+import FromTheCollection from '@/components/prototype/FromTheCollection';
+import BookCard from '@/components/book/BookCard';
+import SignUpCTA from '@/components/auth/SignUpCTA';
+import { type HomeData } from '@/lib/home-data';
+import { HOME_STRINGS, type HomeLang, collectionName } from '@/lib/home-i18n';
+
+// Shared homepage body. The English `/` route renders it with lang="en"; the
+// Spanish `/es` route with lang="es". Keeping a single component means the two
+// language editions can never silently diverge in structure.
+
+export default function HomeView({ data, lang }: { data: HomeData; lang: HomeLang }) {
+  const t = HOME_STRINGS[lang];
+  const { featuredItems, discoverBooks, showcase, counts, collections, blogPosts } = data;
+  const nf = (n: number) => n.toLocaleString(t.locale);
+
+  return (
+    <div className="min-h-screen">
+      <HomePageSchema books={discoverBooks} bookCount={counts.totalBooks} translatedCount={counts.translatedToEnglish} />
+
+      {/* Video Hero */}
+      <HeroSection lang={lang} />
+
+      {/* Collections Grid */}
+      <section id="library" className="bg-gradient-to-b from-[#f6f3ee] to-[#f3ede6] py-16 md:py-24">
+        <div className="px-6 md:px-12 max-w-[1500px] mx-auto">
+          <div className="flex items-baseline justify-between mb-8">
+            <div>
+              <h2 className="text-3xl md:text-4xl text-primary font-display">
+                {t.collectionsHeading}
+              </h2>
+              <p className="text-muted mt-2">
+                <Link href="/search?has_translation=true" className="hover:text-accent-rust transition-colors">{nf(counts.translatedToEnglish)} {t.translationsLabel}</Link>
+                {' '}&middot;{' '}
+                <Link href="/search?first_translation=true" className="hover:text-accent-rust transition-colors">{nf(counts.firstTranslationCount)} {t.firstTimeLabel}</Link>
+                {counts.artworkCount > 0 && (
+                  <>
+                    {' '}&middot;{' '}
+                    <Link href="/artwork" className="hover:text-accent-rust transition-colors">{nf(counts.artworkCount)} {t.artworksLabel}</Link>
+                  </>
+                )}
+                {counts.illustrationCount > 0 && (
+                  <>
+                    {' '}&middot;{' '}
+                    <Link href="/browse/subjects" className="hover:text-accent-rust transition-colors">{nf(counts.illustrationCount)} {t.illustrationsLabel}</Link>
+                  </>
+                )}
+              </p>
+            </div>
+            <Link
+              href="/catalog"
+              className="inline-flex items-center gap-1.5 text-sm font-medium px-4 py-2 rounded-lg border border-accent-rust/30 text-accent-rust hover:bg-accent-rust hover:text-white transition-colors hidden md:flex"
+            >
+              {t.browseCatalog}
+              <span className="text-xs">&rarr;</span>
+            </Link>
+          </div>
+
+          <div className="grid grid-cols-2 lg:grid-cols-4 gap-3 sm:gap-4">
+            {collections.slice(0, 11).map((col, i) => (
+              <Link
+                key={col.slug}
+                href={`/collections/${col.slug}`}
+                className="group relative rounded-xl overflow-hidden aspect-square hover:shadow-lg transition-all hover:-translate-y-0.5"
+              >
+                {col.hero_image ? (
+                  <Image
+                    src={col.hero_image}
+                    alt=""
+                    fill
+                    sizes="(max-width: 640px) 50vw, 25vw"
+                    className="object-cover group-hover:scale-105 transition-transform duration-500"
+                    priority={i < 4}
+                    loading={i < 8 ? 'eager' : 'lazy'}
+                  />
+                ) : (
+                  <div className="absolute inset-0 bg-gradient-to-br from-accent-rust/10 to-accent-gold/10" />
+                )}
+                <div className="absolute inset-0 bg-gradient-to-t from-black/80 via-black/20 to-transparent" />
+                <div className="absolute bottom-0 left-0 right-0 p-4">
+                  <p className="text-white/50 text-xs mb-1 hidden sm:block">
+                    {col.book_count} {t.booksLabel}
+                  </p>
+                  <h3 className="font-serif text-sm sm:text-base lg:text-lg text-white group-hover:text-accent-gold transition-colors line-clamp-2">
+                    {collectionName(lang, col.slug, col.name)}
+                  </h3>
+                </div>
+              </Link>
+            ))}
+            {collections.length > 11 && (
+              <Link
+                href="/collections"
+                className="group relative rounded-xl overflow-hidden aspect-square bg-[#2a2520] flex items-center justify-center hover:shadow-lg transition-all hover:-translate-y-0.5"
+              >
+                <div className="text-center px-4">
+                  <p className="font-serif text-lg sm:text-xl text-white/90 group-hover:text-accent-gold transition-colors">
+                    {t.seeMore(collections.length - 11)}
+                  </p>
+                  <p className="text-white/40 text-xs mt-1">{t.collectionsWord}</p>
+                </div>
+              </Link>
+            )}
+          </div>
+
+          {/* Curated exhibitions link */}
+          <div className="mt-6 flex items-center justify-between">
+            <Link
+              href="/curated"
+              className="group inline-flex items-center gap-2 text-sm text-accent-rust hover:text-accent-rust/80 transition-colors"
+            >
+              {t.curatedExhibitions}
+              <span className="group-hover:translate-x-0.5 transition-transform">&rarr;</span>
+            </Link>
+            <Link
+              href="/collections"
+              className="text-sm text-muted hover:text-accent-rust transition-colors"
+            >
+              {t.allCollections} &rarr;
+            </Link>
+          </div>
+        </div>
+      </section>
+
+      {/* Featured Collection — editorial spread */}
+      {featuredItems.length > 0 && (
+        <EditorialSpread
+          collection={featuredItems[0].collection}
+          books={featuredItems[0].books}
+        />
+      )}
+
+      {/* From the Collection — image-heavy gallery showcase */}
+      <FromTheCollection items={showcase} />
+
+      {/* Discover Section */}
+      <section className="bg-white py-16 md:py-24">
+        <div className="px-6 md:px-12 max-w-[1500px] mx-auto">
+          <h2 className="text-3xl md:text-4xl text-primary mb-3 font-display">
+            {t.discoverHeading}
+          </h2>
+          <p className="text-muted mb-10 max-w-2xl">
+            {t.discoverSubtitle}
+          </p>
+
+          {discoverBooks.length > 0 ? (
+            <div className="grid grid-cols-2 sm:grid-cols-3 md:grid-cols-4 lg:grid-cols-5 gap-4 md:gap-5">
+              {discoverBooks.map((book, i) => (
+                <BookCard key={book.id} book={book} priority={i < 2} />
+              ))}
+            </div>
+          ) : (
+            <div className="text-center py-12">
+              <p className="text-muted mb-4">{t.discoverEmpty}</p>
+              <Link href="/catalog" className="inline-block px-6 py-3 bg-accent-rust text-white rounded-lg hover:bg-accent-rust/90 transition-colors">
+                {t.browseCatalog}
+              </Link>
+            </div>
+          )}
+        </div>
+      </section>
+
+      {/* Blog Section */}
+      <section className="bg-gradient-to-b from-[#f6f3ee] to-[#f3ede6] py-16 md:py-24">
+        <div className="px-6 md:px-12 max-w-[1500px] mx-auto">
+          <div className="flex items-baseline justify-between mb-10">
+            <div>
+              <h2 className="text-3xl md:text-4xl text-primary font-display">
+                {t.blogHeading}
+              </h2>
+              <p className="text-muted mt-2">
+                {t.blogSubtitle}
+              </p>
+            </div>
+            <Link
+              href="/blog"
+              className="text-sm text-accent-rust hover:underline"
+            >
+              {t.blogAllPosts}
+            </Link>
+          </div>
+
+          <div className="grid grid-cols-1 sm:grid-cols-2 lg:grid-cols-4 gap-6">
+            {blogPosts.map((post) => (
+              <Link
+                key={post.slug}
+                href={`/blog/${post.slug}`}
+                className="group bg-white rounded-xl border border-border-light overflow-hidden hover:shadow-lg hover:border-accent-rust/20 transition-[box-shadow,border-color]"
+              >
+                {post.image && (
+                  <div className="aspect-[16/10] relative bg-warm overflow-hidden">
+                    <Image
+                      src={post.image}
+                      alt=""
+                      fill
+                      sizes="(max-width: 640px) 100vw, (max-width: 1024px) 50vw, 25vw"
+                      className="object-cover group-hover:scale-105 transition-transform duration-500"
+                      loading="lazy"
+                    />
+                  </div>
+                )}
+                <div className="p-4">
+                  <span className={`text-xs px-2 py-0.5 rounded-full ${post.tagColor}`}>
+                    {post.tagKey === 'deepDive' ? t.tagDeepDive : t.tagCollection}
+                  </span>
+                  <h3 className="font-display text-lg text-primary mt-2 group-hover:text-accent-rust transition-colors line-clamp-2 leading-snug">
+                    {post.title}
+                  </h3>
+                  <p className="text-sm text-muted mt-1.5 line-clamp-2">
+                    {post.subtitle}
+                  </p>
+                  <p className="text-xs text-faint mt-3">
+                    {post.date} &middot; {post.readTime}
+                  </p>
+                </div>
+              </Link>
+            ))}
+          </div>
+        </div>
+      </section>
+
+      {/* About Section */}
+      <section id="about" className="bg-white py-16 md:py-24">
+        <div className="px-6 md:px-12 max-w-5xl mx-auto">
+          <h2 className="text-3xl md:text-4xl lg:text-5xl text-gray-900 mb-8 leading-tight font-display">
+            {t.aboutHeading}
+          </h2>
+          <div className="space-y-6 text-lg md:text-xl text-gray-600 leading-relaxed">
+            <p>{t.aboutP1}</p>
+            <p>{t.aboutP2}</p>
+            <p className="text-gray-500 text-base">
+              {t.aboutP3Before}
+              <a href="https://embassyofthefreemind.com" className="text-accent-rust hover:underline">{t.efmLinkText}</a>
+              {t.aboutP3After}
+            </p>
+          </div>
+        </div>
+      </section>
+
+      {/* Be part of this */}
+      <section className="py-20 md:py-28" style={{ background: 'var(--bg-dark)' }}>
+        <div className="px-6 md:px-12 max-w-4xl mx-auto">
+          <div className="text-center mb-12">
+            <p
+              className="text-sm uppercase tracking-[0.2em] mb-6"
+              style={{ color: 'var(--accent-gold)' }}
+            >
+              {t.bePartEyebrow}
+            </p>
+            <h2
+              className="text-2xl md:text-3xl lg:text-4xl font-display mb-5 leading-snug"
+              style={{ color: '#f5f0e8' }}
+            >
+              {t.bePartHeading}
+            </h2>
+          </div>
+
+          {/* Support the Library — primary */}
+          <div className="max-w-2xl mx-auto mb-12">
+            <div className="rounded-xl p-8 text-center" style={{ background: 'rgba(255,255,255,0.05)', border: '1px solid rgba(255,255,255,0.08)' }}>
+              <h3 className="text-xl font-display mb-3" style={{ color: '#f5f0e8' }}>
+                {t.supportTitle}
+              </h3>
+              <p className="leading-relaxed mb-6" style={{ color: '#a09a90' }}>
+                {t.supportBody}
+              </p>
+              <div className="flex flex-col sm:flex-row gap-3 justify-center">
+                <Link
+                  href="/support"
+                  className="inline-flex items-center justify-center gap-2 px-8 py-3.5 rounded-full text-sm font-medium transition-all hover:brightness-110"
+                  style={{ background: 'var(--accent-rust)', color: '#fff' }}
+                >
+                  {t.howToSupport}
+                </Link>
+                <Link
+                  href="/auth/signin"
+                  className="inline-flex items-center justify-center gap-2 px-8 py-3.5 rounded-full text-sm font-medium transition-all hover:brightness-110"
+                  style={{ background: 'rgba(255,255,255,0.08)', color: '#a09a90' }}
+                >
+                  {t.createAccount}
+                </Link>
+              </div>
+            </div>
+          </div>
+
+          {/* Other ways to participate */}
+          <div className="grid grid-cols-1 sm:grid-cols-2 gap-6 max-w-2xl mx-auto">
+            <Link
+              href="/contribute"
+              className="rounded-lg p-5 text-center hover:brightness-110 transition-all"
+              style={{ background: 'rgba(255,255,255,0.03)', border: '1px solid rgba(255,255,255,0.06)' }}
+            >
+              <p className="text-sm font-medium mb-1" style={{ color: '#f5f0e8' }}>{t.contribute}</p>
+              <p className="text-xs leading-relaxed" style={{ color: '#6b6560' }}>
+                {t.contributeDesc}
+              </p>
+            </Link>
+            <Link
+              href="/developers"
+              className="rounded-lg p-5 text-center hover:brightness-110 transition-all"
+              style={{ background: 'rgba(255,255,255,0.03)', border: '1px solid rgba(255,255,255,0.06)' }}
+            >
+              <p className="text-sm font-medium mb-1" style={{ color: '#f5f0e8' }}>{t.developers}</p>
+              <p className="text-xs leading-relaxed" style={{ color: '#6b6560' }}>
+                {t.developersDesc}
+              </p>
+            </Link>
+          </div>
+
+          {/* Free account nudge — for anonymous users only */}
+          <SignUpCTA variant="inline" />
+        </div>
+      </section>
+
+      {/* Search */}
+      <section className="bg-[#f6f3ee] py-16 md:py-20">
+        <div className="px-6 md:px-12 max-w-2xl mx-auto text-center">
+          <h2 className="text-2xl md:text-3xl font-display text-stone-900 mb-3">
+            {t.searchHeading}
+          </h2>
+          <p className="text-stone-500 text-sm mb-8">
+            {t.searchStats(nf(counts.totalBooks), nf(counts.authorCount), counts.languageCount)}
+          </p>
+          <form action="/search" method="get" className="relative max-w-lg mx-auto mb-6">
+            <svg className="absolute left-4 top-1/2 -translate-y-1/2 w-5 h-5 text-stone-400" fill="none" stroke="currentColor" viewBox="0 0 24 24">
+              <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M21 21l-6-6m2-5a7 7 0 11-14 0 7 7 0 0114 0z" />
+            </svg>
+            <input
+              type="text"
+              name="q"
+              placeholder={t.searchPlaceholder}
+              className="w-full pl-12 pr-12 py-3.5 bg-white border border-stone-200 rounded-full text-stone-900 placeholder-stone-400 focus:outline-none focus:ring-2 focus:ring-accent-rust/20 focus:border-accent-rust shadow-sm"
+            />
+          </form>
+          <p className="text-stone-400 text-sm">
+            {t.browseBy}{' '}
+            <Link href="/browse/titles/A" className="text-stone-600 hover:text-accent-rust transition-colors underline underline-offset-2">{t.byTitle}</Link>
+            {' '}&middot;{' '}
+            <Link href="/browse/authors/A" className="text-stone-600 hover:text-accent-rust transition-colors underline underline-offset-2">{t.byAuthor}</Link>
+            {' '}&middot;{' '}
+            <Link href="/browse/years/1500s" className="text-stone-600 hover:text-accent-rust transition-colors underline underline-offset-2">{t.byYear}</Link>
+            {' '}&middot;{' '}
+            <Link href="/gallery" className="text-stone-600 hover:text-accent-rust transition-colors underline underline-offset-2">{t.byImages}</Link>
+          </p>
+        </div>
+      </section>
+
+      {/* Footer */}
+      <footer className="bg-gradient-to-b from-[#f6f3ee] to-[#f3ede6] py-16 md:py-24">
+        <div className="px-6 md:px-12 max-w-5xl mx-auto">
+          <div className="max-w-4xl border-t border-stone-300 pt-10 mt-8">
+            <p className="text-sm uppercase tracking-[0.2em] text-stone-500 mb-6">
+              {t.inSpiritOf}
+            </p>
+            <div className="flex flex-col md:flex-row md:items-start gap-8 md:gap-12">
+              <div className="flex-1">
+                <h3 className="text-xl md:text-2xl text-stone-800 mb-1 font-display">
+                  Marsilio Ficino
+                </h3>
+                <p className="text-stone-500 text-sm mb-3">
+                  {t.ficinoRole}
+                </p>
+                <p className="text-stone-600 text-base leading-relaxed">
+                  {t.ficinoBio}
+                </p>
+              </div>
+              <div className="flex-1">
+                <h3 className="text-xl md:text-2xl text-stone-800 mb-1 font-display">
+                  Cosimo de&apos; Medici
+                </h3>
+                <p className="text-stone-500 text-sm mb-3">
+                  {t.cosimoRole}
+                </p>
+                <p className="text-stone-600 text-base leading-relaxed">
+                  {t.cosimoBio}
+                </p>
+              </div>
+            </div>
+            <div className="mt-8 bg-accent-gold/5 rounded-lg p-5 border border-accent-gold/15">
+              <p className="text-stone-700 text-base leading-relaxed">
+                <strong>{t.closingStrong}</strong>{t.closingRest}
+              </p>
+            </div>
+          </div>
+        </div>
+      </footer>
+    </div>
+  );
+}
